refactor(ShowMain): lib,ui,model
create the `lib`,`ui`,`model` for ShowMain
